Fixed problems in the 2nd task
