+++ task_02/src/max_element.cpp
@@ -1,12 +1,12 @@
-#include "max_element.hpp"
+// #include "max_element.hpp"
 
-int get_max_element(int* array, int size)
-{
-    int max_element = array[0];
-    for (int i{ 0 }; i < size; i++)
-    {
-        if (max_element < array[i])
-            max_element = array[i];
-    }
-    return max_element;
-};
+// int get_max_element(int* array, int size)
+// {
+//     int max_element = array[0];
+//     for (int i{ 0 }; i < size; i++)
+//     {
+//         if (max_element < array[i])
+//             max_element = array[i];
+//     }
+//     return max_element;
+// };
