+++ task_02/src/max_element.hpp
@@ -1,1 +1,10 @@
-int get_max_element(int* arr[] , int len);
+int get_max_element(int* array, int size)
+{
+    int max_element = array[0];
+    for (int i{ 0 }; i < size; i++)
+    {
+        if (max_element < array[i])
+            max_element = array[i];
+    }
+    return max_element;
+};
